cambios 4

+++ proyecto2obligatorio/src/main/java/com/example/Cafetera.java
@@ -3,11 +3,11 @@ import java.util.concurrent.Semaphore;
 
 public class Cafetera {
     Semaphore semaforo = new Semaphore(1);
-    long tiempoUsada = 0;
+    double tiempoUsada = 0;
 
     public void usar(Pedido pedido) throws InterruptedException {
         semaforo.acquire();
-        long inicio = System.currentTimeMillis();
+        double inicio = System.currentTimeMillis();
 
         try {
             System.out.println("Pedido " + pedido.numero + " usa cafetera");
